+++ index.html
@@ -57,6 +57,7 @@
         <div class="wall-left"></div>
         <div class="wall-right"></div>
     </div>
-    <script src="./js/main.js" type="module"></script>
+
+<script src="./js/main.js" type="module"></script>
 </body>
 </html>
