--- js/createDate.js
@@ -1,5 +1,0 @@
-//функция создание даты 
-export const createDate=()=>{
-    const date =` ${new Date().toLocaleDateString()} ${new Date().toLocaleTimeString()}`
-    return date
-}
