+++ js/enemyAttack.js
@@ -1,4 +1,4 @@
-import{getRandom} from './getRandom.js';
+
 
 export const ATTACK = ['head', 'body', 'foot'];
 
@@ -10,12 +10,3 @@ export const HIT = {
 
 
 //создание атаки врага
-export const  enemyAttack=()=> {
-    let defence = ATTACK[getRandom(3)-1];
-    let hit = ATTACK[getRandom(3)-1];
-    return {
-        value: getRandom(HIT[hit]),
-        hit: hit,
-        defence,
-    }
-}
